Fix resizing static maps removing doors (#42)
Static maps, when resized, cleared the map then
replaced each tile in turn. However, there were
validations for placing door tiles, to make sure its
entrances were clear (floor or room tile). The
clearing of the map was done by placing NOTHING
tiles, which did not allow doors to be placed.

The fix is to clear the map with FLOOR tiles instead.