Fix co-op AI getting stuck in weapon switch loop
Player special used flag is only cleared on two
consecutive no-commands, but if AI attempts to
switch weapon before this happens, it will get in
a loop of attempting to switch weapons without
anything happening. Instead, the special used
flag is replaced with a check on whether the last
command was special (button 2 + direction).