Add game wait state (#405)
Fixes exit area sometimes showing for
networked PVP games. There is a single
frame where the client has joined but the
players haven't spawned, causing the game
to incorrectly think that PVP is over and
one player has won.

Also correctly shows the
"Waiting for players" message if a PVP
mode is played but only one player is
present.