Fix custom ammo not being picked up (fixes #407)
Custom data was only being unloaded on the next map load… for new maps.
If an older map was loaded, the previous custom data is still present.
This would normally not be noticed, since the old map won’t use the
custom data anyway.
HOWEVER, if the custom data contains duplicates of the normal map data,
such as ammo types, the game will spawn those ammo pickups instead of
the normal ammo types.
The duplicates have exactly the same names and properties but are
subtly different, i.e. different IDs.
This means the ammo cannot be picked up, but is otherwise a duck.