Fix name of variables and truncate distance
The name of variables were modified during the search/replace. This fix
the problem. The variables of distance now is truncated to the nearest
lower integer (same behavior as previously). This is needed because the
vector function to measure distance returns a floating-point, but the
game logic uses integer. Explicit type conversion using `(int)` results
in the same previous behavior.