Swap guns on pick up (fixes #513)
When picking up guns when the player has no empty slots, drop the current gun that the player has.

Also, allow picking up the same gun even when the player has that gun already - instead the gun will be dropped again, and ammo (if low) will be replenished up to the default amount.