Highlight access level in editor; don't spawn keys in doors (fixes #619)
Keys can still be accidentally spawned on doors if for some reason they were open while the map was being generated. When the door is then closed, this could cause the key to be inaccessible. Instead, mark walls of locked rooms as locked as well, which prevents keys from spawning on them.